requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
python-docx>=0.8.11
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
nltk>=3.8.1
spacy>=3.6.0
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
altair>=4.0.0
sqlite3
